packages/api-gateway/pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
description = "API Gateway"
authors = [
    {name = "Milan Amsterdam",email = "milan@amsterdam.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
httpx = "^0.28.1"

[tool.pytest.ini_options]
pythonpath = ["src"]

